requirements.txt
ultralytics==8.3.183
fastapi==0.111.0
uvicorn[standard]==0.30.3
pillow==10.4.0
numpy==1.26.4
google-generativeai==0.7.2
python-multipart==0.0.9
onnxruntime==1.18.1   # <- important for ONNX on CPU